update 20:49
LOG:
completed new UI
removed "run_string_ex", invalid sig

+++ include/cheats/Visuals.h
@@ -119,7 +119,7 @@ namespace raicu::cheats {
 		};
 
 	private:
-		static void DrawCrosshair(float width, float length, float offset, ImU32 color, float rounding, bool outline, ImU32 outlineColor, float outlineThickness);
+		static void DrawCrosshair(int width, int length, int offset, ImU32 color, float rounding, bool outline, ImU32 outlineColor, float outlineThickness);
 		static void DrawAimbotFOV(float FOV);
 
 		static void Drawbox(int CurrentEnt);
